+++ app/ide-desktop/lib/content/src/index.ts
@@ -125,11 +125,11 @@ class Main {
                     }
                     /** This package is an Electron desktop app (i.e., not in the Cloud), so
                      * we're running on the desktop. */
-                    const platform = authentication.Platform.desktop;
+                    const platform = authentication.Platform.desktop
                     const onAuthenticated = () => {
                         hideAuth()
                         appInstance.run()
-                    };
+                    }
                     authentication.run(logger, platform, onAuthenticated)
                 } else {
                     appInstance.run()