+++ app/ide-desktop/lib/dashboard/craco.config.ts
@@ -1,14 +1,14 @@
 /** @file Configuration file for the `craco` package.
- * 
+ *
  * `craco` is used in place of the default `react-scripts` package provided by `create-react-app`.
  * `craco` is a wrapper around `react-scripts` that allows us to customize the build process. We use
  * it to disable the `ModuleScopePlugin`. The `ModuleScopePlugin` prevents importing modules outside
  * of the `src` directory. This package imports files from sibling packages in the monorepo, so the
  * `ModuleScopePlugin` must be disabled.
- * 
+ *
  * See: https://stackoverflow.com/a/60353355. */
 
-import * as craco from '@craco/types';
+import * as craco from '@craco/types'
 import * as webpack from 'webpack'
 
 // =================
@@ -16,7 +16,7 @@ import * as webpack from 'webpack'
 // =================
 
 /** Name of the `ModuleScopePlugin` plugin. */
-const MODULE_SCOPE_PLUGIN = 'ModuleScopePlugin';
+const MODULE_SCOPE_PLUGIN = 'ModuleScopePlugin'
 
 // ==============================
 // === Disable Plugin By Name ===
@@ -25,16 +25,16 @@ const MODULE_SCOPE_PLUGIN = 'ModuleScopePlugin';
 /** Finds and disables the plugin with the given `name` in the Webpack configuration, if the plugin
  * exists. */
 const disablePluginByName = (webpackConfig: webpack.Configuration, name: string): void => {
-    const plugins = webpackConfig.resolve?.plugins;
+    const plugins = webpackConfig.resolve?.plugins
     if (!plugins) {
-        return;
+        return
     }
 
     const pluginIndex = plugins.findIndex(
         ({ constructor }) => constructor && constructor.name === name
-    );
+    )
     if (pluginIndex !== -1) {
-        plugins.splice(pluginIndex, 1);
+        plugins.splice(pluginIndex, 1)
     }
 }
 
@@ -46,7 +46,7 @@ module.exports = {
     webpack: {
         configure: webpackConfig => {
             disablePluginByName(webpackConfig, MODULE_SCOPE_PLUGIN)
-            return webpackConfig;
-        }
-    }
-} as craco.CracoConfig;
+            return webpackConfig
+        },
+    },
+} as craco.CracoConfig