+++ app/ide-desktop/lib/dashboard/src/authentication/src/authentication/config.ts
@@ -31,34 +31,38 @@ export const OAUTH_RESPONSE_TYPE: OAuthResponseType = 'code' as OAuthResponseTyp
 
 /** The AWS region in which our Cognito pool is located. This is always set to `eu-west-1` because
  * that is the only region in which our Cognito pools are currently available in. */
-type AwsRegion = 'eu-west-1' & { _brand: "AwsRegion"; }
+type AwsRegion = 'eu-west-1' & { _brand: 'AwsRegion' }
 /** ID of the "Cognito user pool" that contains authentication & identity data of our users.
  *
  * This is created automatically by our Terraform scripts when the backend infrastructure is
  * created. Look in the `enso-org/cloud-v2` repo for details. */
-type UserPoolId = string & { _brand: "UserPoolId"; }
+type UserPoolId = string & { _brand: 'UserPoolId' }
 /** ID of an OAuth client authorized to interact with the Cognito user pool specified by the
  * {@link UserPoolId}.
  *
  * This is created automatically by our Terraform scripts when the backend infrastructure is
  * created. Look in the `enso-org/cloud-v2` repo for details. */
-type UserPoolWebClientId = string  & { _brand: "UserPoolWebClientId"; }
+type UserPoolWebClientId = string & { _brand: 'UserPoolWebClientId' }
 /** Domain of the Cognito user pool used for authenticating/identifying the user.
  *
  * This must correspond to the public-facing domain name of the Cognito pool identified by the
  * {@link UserPoolId}, and must not contain an HTTP scheme, or a pathname. */
-type OAuthDomain = string  & { _brand: "OAuthDomain"; }
+type OAuthDomain = string & { _brand: 'OAuthDomain' }
 /** Possible OAuth scopes to request from the federated identity provider during OAuth sign-in. */
-type OAuthScope = ('email' | 'openid') & { _brand: "OAuthScope"; }
+type OAuthScope = ('email' | 'openid') & { _brand: 'OAuthScope' }
 /** The response type used to complete the OAuth flow. "code" means that the federated identity
  * provider will return an authorization code that can be exchanged for an access token. The
  * authorization code will be provided as a query parameter of the redirect URL. */
-type OAuthResponseType = 'code'  & { _brand: "OAuthResponseType"; }
+type OAuthResponseType = 'code' & { _brand: 'OAuthResponseType' }
 /** The URL used as a redirect (minus query parameters like `code` which get appended later), once
  * an OAuth flow (e.g., sign-in or sign-out) has completed. These must match the values set in the
  * Cognito pool and during the creation of the OAuth client. See the `enso-org/cloud-v2` repo for
  * details. */
-export type OAuthRedirect = (typeof DESKTOP_REDIRECT | 'http://localhost:8081' | 'https://cloud.enso.org')  & { _brand: "OAuthRedirect"; }
+export type OAuthRedirect = (
+    | typeof DESKTOP_REDIRECT
+    | 'http://localhost:8081'
+    | 'https://cloud.enso.org'
+) & { _brand: 'OAuthRedirect' }
 /** Callback used to open URLs for the OAuth flow. This is only used in the desktop app (i.e., not in
  * the cloud). This is because in the cloud we just keep the user in their browser, but in the app
  * we want to open OAuth URLs in the system browser. This is because the user can't be expected to