+++ app/ide-desktop/lib/dashboard/src/authentication/src/http.tsx
@@ -135,4 +135,3 @@ export class Client {
     return new HttpResponse(response);
   };
 }
-