+++ app/ide-desktop/lib/dashboard/src/authentication/tailwind.config.ts
@@ -1,6 +1,6 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-    content: ["./src/**/*.tsx"],
+    content: ['./src/**/*.tsx'],
     theme: {
         extend: {},
     },