+++ CHANGELOG.md
@@ -111,6 +111,7 @@
 - [Named arguments syntax is now recognized in IDE][5774]. Connections to
   function arguments will now use named argument syntax instead of inserting
   wildcards on all preceding arguments.
+- [Added boilerplate React app for authorization via Cognito+AWS Amplify][5798].
 
 #### EnsoGL (rendering engine)
 
@@ -517,6 +518,7 @@
 [5774]: https://github.com/enso-org/enso/pull/5774
 [5779]: https://github.com/enso-org/enso/pull/5779
 [5785]: https://github.com/enso-org/enso/pull/5785
+[5798]: https://github.com/enso-org/enso/pull/5798
 [5802]: https://github.com/enso-org/enso/pull/5802
 [5850]: https://github.com/enso-org/enso/pull/5850
 [5863]: https://github.com/enso-org/enso/pull/5863